refactor: Refacotring the insertion operation
